modify request config
