--- crypto_arbitrage_detector/configs/strategy_config.py
@@ -1,5 +1,0 @@
-DEFAULT_SLIPPAGE_BPS = 100  # Default slippage in basis points (1% slippage)
-DEFAULT_TX_AMOUNT = 1000000  # Default transaction amount in lamports (1 SOL = 1,000,000 lamports)
-SOL_MINT = "So11111111111111111111111111111111111111112"  # Default SOL mint
-SCRAPER_API_KEY = "962b799f81dbe3b72467b62db9474331"
-PROXY_URL = f"http://scraperapi:{SCRAPER_API_KEY}@proxy-server.scraperapi.com:8001"
